refactor(ipc): split the conversation draft-filing handlers into their own registrar (#1623)
The deferred follow-up from #1623 (arch D4): register-conversation.ts still
carried the ~470-line draft-filing cluster — the Approve-on-a-draft-card
handlers for propose_notes / propose_sources / set_properties / propose_claims,
the note-refactor/reorg/delete/body drafts, and propose_compute.

- New `register-conversation-drafts.ts` with `registerConversationDrafts()`
  holds those handlers plus the three draft-only helpers they use
  (conversationProvenance, ensureDraftItems, buildClaimNoteContent). Each still
  files through the approval engine and auto-approves the already-reviewed card
  — the Trust Principle is unchanged (pure relocation, byte-identical bodies).
- Wired alongside registerConversation() in ipc.ts, mirroring the proposals
  split. register-conversation.ts now holds just the conversation lifecycle +
  the SEND streaming/abort/retry surface: 954 → 399 lines.

The #1612 handler test now registers both registrars; registration snapshot
unchanged (same channels). ipc (16) + llm/trust (316) suites green; lint clean.
